[xs] merge auth_helper with auth.annotation
There were two helper methods in auth_helper. But i can easily be merged with annotation
 as these annotations are the major consumer of these two methods. And it will reduce
 the number of files, which is always good.

Ref - #260
